Remove hyperparameter notice
Due to a runtime error I had to update the space to pin pytorch geometric to 2.1.0 and also updated the space to run with the same default options as described here, so I think this notice can be removed. I keep the number of samples at 10 for now due to runtime constraints but people can set it higher if they want. 

I think the colab notebook also needs to fix the dependency, it threw a few errors when I tried it earlier: `pip install git+https://github.com/pyg-team/pytorch_geometric.git@2.1.0
`